Bug 701083(2): Write the new key/value to DecodeParms
There was a lengthy block of code that handled PDFs with buggy DCTDecode
dictionaries with a missing, or incorrect Height entry. Since we now have a
second key/value needing added to the DecodeParms and in two places, I've
split that into a more general procedure, so we can add multiple new keys in
one step.

This is used to add/patch the Height key, and add the new ErrsAsEOD key
(telling the CCITT fax decode filter to treat an error as a EOD).