Fix RunLengthDecode filter
We were treating the RunLengthDecode filter as a 'simple' filter,
because the PDF Reference says it has  no parameters.

This is true for *PDF* but it isn't true for PostScript. Since ours is
a PostScript implementation we do actually need to set the defaults.

Because setting the defaults involves using a stream template we can't
use the simple filter form, and need a new filter to handle RunLength
decoding.

tests_private__pdf__PDFIA1.7_SUBSET__CATX2905
tests_private__comparefiles__Bug693000

show progressions with this in place.