Graphics library - match gsaves with grestores
Auditing as part of the pdfi work, when the graphics library executes
a gs_gsave() it should match it with a gs_grestore(), even in the case
of an error.

In addition, if gs_setcolorspace failed in image_PaintProc() it would
leak the pattern enumerator. Because this is only called from PCL
the memory is not garbage collected.