Move code to retrieve rects from PDF Arrays into pdf_array.c
There are (or will be) many places we want to turn a PDF array into
a gs_rect (annotations, fonts etc). So move the static code from
pdf_pattern into pdf_array and make it non-static so other clients
can use it.

Also move the code for turning an array into a gs_matrix.