Modify the new gp_file API so that gp_fread etc match fread
The new gp_fread() and gp_fwrite() have a parameter list which (more or
less) matches the list for the C run-time functions fread() and fwrite()

However, the return values do not match the returned values from the
C runtime functions. This is because, internally, the gp_file structure
has function pointers which only take 3 parameters instead of 4. This
is handled by multiplying the size of the record by the number of records
to be read/written, and then reading that many bytes.

Since the underlying functions are (at this time) the C runtime functions
this means that th returned value becomes the number of bytes read
(or written) instead of the number of *records* read/written.

This commit modifies that internal API so that the gp_file member
functions take 4 parameters (thus matching both gp_fread and fread)
and pass them to the underlying functions. The return value is then
the number of records, as expected.

Should we ever encounter an implementation which doesn't match the
C runtime fread() we could implement it by multiplying the size by the
count (as was done prior to this commit) and then, assuming the return
value was the number of bytes read, dividing the return value by the
size, in order to retrieve the number of records read.