PDF interpreter - don't execute newpath when ending text blocks
Bug 701178 - "PDF missing two bounding boxes on process"

The missing rectangles are defind in the content stream between a
number of BT/ET text blocks, and only finally stroked after all the
text has been emitted.

Because we execute a newpath when ending a text block (to avoid dangling
path elements from text), and prior path elements are lost.

This commit only executes a newpath when the text rendering mode is one
of the 'clip' modes, as that is handled differently, and we need the
newpath.

Its a pretty poor way to output the paths, and I suspect that it would
likely run into trouble with the text clipping modes, but since those
aren't used here its hard to tell.