Make .pdfexectoken not an operator
In a previously round of security fixes, I made .pdfexectoken an operator. That
turns out to cause problems because we (normally) attempt to recover from errors
in PDF files, so we don't want .pdfexectoken to cleanup the stacks in the event
of an error.

In addition, the original problem this solved is solved in a more robust way
(executeonly on the transient procedures).