Fix spot colour counting
Two oversights; a Pattern colour space can contain a Shading dictionary
and in that case we need to check the Shading dictionary for spot colours

A ExtGState can have an SMask and the SMask can have a /G which can
refer to an XObject, we need to check the XObject and all its Resources
for spot colours too.

Note, the routines pdfi_Shading_dict() and pdfi_check_XObject_dict()
are intended to be handed a dictionary of Resources of the relevant
type, the result of traversing a Resources dictionary. To facilitate
checking an individual XObject or Shading dictionary I've moved the
code out of those routines and into separate routines.

Spotted and fixed 2 memory leaks while refactoring this.