Extension and tidying up of the resource checking
The Shading code was missing some countdowns.

We could potentially short-circuit some work in ExtGState processing
if we don't have a spot-capable device.

Moved some code out into separate routines so that we can check an
instance of a resource type separately to checking a dictionary full
of particular resource type definitions.

Fixed a missing countdown on annotation processing.

Fixed some reference counting problems with the page dictionary.

Added a load of comments, fixed some others.

Addressed some compiler warnings.

I had hoped to make the code which checks a dictionary of a particular
Resource type general, but unfortunately the ColorSpace checks are
slightly different to all the others. I may revisit this in future. I'm
slightly annoyed by having so many almost-but-not-quite-identical
routines.

Upgraded the cache cleanup and consistency checking. There was a problem
here in that a recently used object in the cache might also be referenced
by a composite object (dictionary or array) which is *also* stored in
the cache. This can legitimately mean that the recently used object has
a reference count > 1. Of course, if we count down the cache reference
to the early object, and then later free the referencing object, then
the early object should be counted down and freed at that point.

I've removed the debug printout from the code, and added a #ifdef DEBUG
section instead which walks over the cache multiple times. On each pass
it counts down and frees any object with a reference count of 1 (only
referenced by the cache). If it does not count down any objects it exits
the loop. If any objects remain in the cache it prints a report of the
remaining objects.

With my test file this previously reported 4 objects having a reference
count of > 1 on exit, with this code the cache exits cleanly and reports
no problems.