PDF interpreter - improve repair of broken PDF files
If a file had a startxref token which was not followed by any further
actual tokens (eg nothing but % comment lines) then the repair code
failed to account for this, and ended up trying to pop non-existent
objects from the stack.

This is because the code did not test the result of the 'token' operator
to ensure that a token had been read, it assumed one would be. In this
commit we check the result and, if its 'false' (no token read) we don't
attempt to pop the non-existent token from the stack, but instead follow
the existing error path and resort to scanning the PDF file for the
trailer dictionary.

In addition to fixing the customer reported file, this also fixes the
existing test suite file

tests_private/pdf/sumatra/repairing_leaks_object.pdf