"Fix" contrib and test uses of gp_fopen.
Here we fix a lot of contrib devices that had been broken by the
changes to gp_file. I believe they all should work OK now.

One of the routines was calling lib_fopen in an unsupported way; this
can't easily be fixed, so we just call gp_fopen instead now.

We also "fix" some test code that was calling gp_fopen and then doing
things that we can't easily do with a gp_file *. This code has reverted
to using FILE *'s and fopen - this might cause issues on compilation as
this code is untested, but it's better than compiling to give code that
crashes.