Fix Indexed 'install' procedure
The Install procedure for Indexed spaces was calling the 'install'
procedure for its base space. This is incorrect. Colour spaces are
supposed to be set from the lowest (most 'base') to the highest. So the
Indexed space's base space should already be set, and installed. Trying
to set it again is a bad thing and can cause Seg Faults.

However, the PCL interpreter relies on the Indexed space doing this,
contrary to the expected behaviour it does not set the base space before
setting the Indexed space, and relies on the Indexed space to 'install'
the base space.

Setting the base space caused unexpected (and inexplicable) differences
in halftoned output. I 'suspect' that this is because it sets the three
initial components of pgs->color[0].ccolor, whereas the previous code
only resulted in setting the initial value.

So instead, call the base space's 'install' procedure directly, this
doesn't set the colour values (which gs_setcolorspace does) and results
in no differences.

For me this resolves the crashes that were being seen in the 32-bit
builds of Ghostscript and shows no differences.