Improve -dFILTERIMAGE
Previously if we were supplied too much image data, we could read off
the end of the supplied data and consume input that should have been
interpreted.

This limits the consumption of data so that we don't read more than we
expected.

NB! We still don't support type 3 images, as that requires us to process
2 images at once.