Preparatory work for the PDF interpreter in C
Currently the pdfwrite device assumes that a font with FontType 3
(ft_user_defined) *will* require an exit to the interpreter. The
CharProc capture is set up so that this happens in a complicated dance.

When we have the PDF interpreted in C we do not need to exit to the
interpreter, and so we need to capture the CharProc in the same way we
do for PCL 'user defined' fonts (eg the HPGL stick font and PCL bitmap
fonts).

Obviously to be able to do this we need to be able to differentiate
between a PostScript type 3 font (or a type 3 font built by the PostScript
PDF interpreter) and a type 3 font built by the C PDF interpreter. So
the first thing we do is add a new ft_PDF_user_defined font tyep.

Then update the graphics library so that it treats the PDF type 3 font
in exactly the same way as it does a PostScript type 3 font.

In the pdfwrite device, in general we want to treat a PDF type 3 font
as being the same as a PS type 3.

Finally, in pdf_text_process() the bulk of the real work. If we have a
ft_PDF_user_defined font, then set 'early_accumulator' which means we
push the accumulator early, because we won't need to exit to the
interpreter. This means that in the 'early_accumulator' branch we need
to account for the possibility that the text operation won't be a
'TEXT_FROM_CHARS' (it always is for PCL, which is why this didn't need
to be handled before). There was also some bad error handling in here,
we do a gs_gsave() which is undone by a gs_grestore() in comp[ete_charproc()
but if we had an error before that we would not undo the gs_gsave whhich
could lead to ridiculous levels of nested gsaves.