pdfwrite - fix an error path
This turned up while working on the PDF interpreter in C. Previously we
had no examples which exercised this error pathway, because it was only
used by PCL and none of our PCL examples threw an error here.

If we've started accumulating a CharProc, then we have started a pdf
substream. If we get an error, then we need to terminate that substream
before returning (the caller can't know that the substream was begun).