Many improvements in font/text/text state processing
Implement the various text state operators

Add pdfi_gsave/grestore around pages, this is to ensure that the current
font counting is correct, and that we don't end up with a dangling
reference from the gstate at the end of the job. NOTE! when executing
gsave/grestore in the interpreter, if there is any possibility of the
current font being changed between the gsave and grestore, we must now
use pdfi_gsave() and pdfi_grestore() so that the font reference counting
works properly. If in doubt use the pdf_ versions.

Fix font selection when we have found an existing built font to use,
previously we were throwing an error instead of using the newly found
font.

Add REFCNT_DEBUG to type 3 font creation (allocating bytes directly for
a structure doesn't do this for us)

countdown the references in the pdf_font_type 3 structure to CharProcs
Resources and Encoding, to fix memory leaks.

countup Encoding array referenced from pdf_font_type3 when we create it,
otherwise the reference count is always wrong.

Pop the BT keyword from the stack when we encoutner it. Otherwise we end
up corrupting the stack.