Resolve font_dir confusion
We were allocating two font_dir structures, one of which we were storing
in the PDF contexrt, the other in the 'gs_lib_ctx. I'm not sure why (or
indeed if) we need a gs_lib_ctx. It seems that we have control, for now
at least, over the fotn_dir which is stored in the font structure and
its that which the gtaphics library seems to use.

So avoid creating a gs_lib_ctx->font_dir, only purge and free the one
in the PDF context.