Rework ROP transparency
Rework ROP behaviour with transparency based upon the results
from roptest.pxl seen on an HP printer.

The old style ROP transparency code used to treat white pixels
as special. The new style code does everything in boolean logic
by modifying the ROP that is used when S or T pixels are set to
be transparent.

This enables us to rip out several cases from the ROP
processing loops, as we never need to check for transparent
pixel values anymore.