Break out the stream 'cleanup' code into separate routines
The code for saving the current state of various parameters (eg the
current file position, gsave depth, transparency group countin etc) in
a local variable, forcing the state back to the saved state after
interpreting a stream, and returning the state to the locally saved
copy is used in two places, and is getting extensive. So break the
save/restore/cleanup code into separate functions.

This should improve readability, make it easier to keep the functions
in sync if we add more state variables and ensure we do the same work
in both places where this is currently called.