replace gs_gsave and gs_grestore with pdfi_gsave and pdfi_grestore
For the benefit of tracking fonts, we need to use the pdfi_ variants.

This is complicated by the fact that setcachedevice does a gsave which
is not grestore'd until after a CharProc is completed, see the comments
in pdf_font.c, pdf_d0().

pdf_pattern.c still needs to be addressed, there are complications in
that file and attempting a naive change resulted in many seg faults.