Use pdfi_gsave/grestore instead of gsave/grestore for patterns
The surprise in this case is that gs_copygstate also copies the gsave
nesting level. So if we copy an old gstate over one at a greater depth,
the gsave level is incorrect. This causes trouble when we do a
pdfi_gsave/pdfi_setgstate/pdfi_grestore. The nesting level is incorrect
when we get to pdfi_grestore, its earlier than the save level enclosing
the stream in fact, so we don't execute the grestore.

Fix that by patching up the gsave level when we do gs_copygstate.

Add pdfi_grestore() on a few error paths that were not previously
restoring the gstate.