Fix buffer overrun in mswinpr2 device
Bug #701281 "ACTIONABLE_HEAP_CORRUPTION when try print with -sOutputFile"

The mswinprs2 device uses the very old 16-bit GetProfileStringW() API
call to get the 'Devices' in order to look for a specified printer.

The code however assumes that the buffer length passed is in bytes, and
in fact it should be in wchar_. This means that a sufficiently large
Devices section could overrun the buffer, as it thinks there are more
wchar_t available than there are.

in addition, if the section is very large we could run out of space in
the string altogether, and then would be unable to find the designated
printer.

This commit alters the code so that the length passed to GetProfileStringW
is in wchar_t and also examines the returned length. If the returned
length indicates possible truncation of the return, then we free the
buffer, increase its size by 4Kb and try again. Repeatedly until we
either exhaust memory or get a large enough buffer.