Fix memory leak related to transparency mode, pdf14 device
The pdf14 device lives forever once it is created, and it also holds a
reference to the base device (in its -->target).  So both of these
devices were not getting deallocated, causing a memory leak.

This fix keeps a pointer to the pdf14 device so it can dereference it in
the context cleanup.

Note that xps suffers from this same bug (see Bug 701284).
There may be issues with running pdfi inside gpdl or gs, due to garbage
collection.  The whole pdf14/transparency device memory management may
need to be revisted.

Test file: tests/pdf/GridDownscaleTest0.pdf