Bug 701284: Fix leak of pdf14device in xps.
The pdf14device is created and then "retained". This is required
for gs, because postscript can call "currentdevice" and get a
non reference counted pointer to a device. It is cleared up
correctly because of gc.

Other languages have no gc, and so leak. The fix is to enable
those languages to get the pdf14device "unretained". We add an
graphics lib call to allow this to happen.

Transparent Patterns also create pdf14devices, which are again
retained, despite not needing to be. In gs this causes no problems
as they are cleared by the restore, but in other languages they
leak. We make them unretained here, with no ill effects.