Cleanup interpreter objects for Patterns when complete
First up, when creating a type 2 (smooth shading) pattern, we don't
need to create the interpreter pattern 'context', because we don't need
to run a pattern PaintProc. So don't create one when we don't need one.

Managing the lifetime of a pattern is unfortunately complicated. The
pattern is created when we set the colour space and although it is
reference counted, and has a client_data pointer, there is no way for
the interpreter to be called when the pattern is replaced and discarded
so there's no obvious way for us to clean up our allocation.

I did look at gs_notify_register, which calls a 'notification' procedure
when an object is discarded. But that only works for fonts, adding it to
patterns would be complicated and in any event, if we are going to do
that amount of work it would be better to do the notification in the
memory manager so that it works for all kinds of objects.

So instead; modify pdfi_grestore so that if we are grestore'ing and
either the current stroke or fill colour is a pattern, and the pattern
reference count is 1, call the pattern cleanup code. Modify our colour
handling so that whenever we replace the current colour space we also
check to see if its a pattern, and if its reference count is 1, and if
so clean it up. Both these cases *should* mean that the pattern space
will be counted down and discarded, so we want to clean up our memory
too.

I did this by changing calls like gs_setgray() and gs_setcolorspace() to
pdfi_gs_setgray() and pdfi_gs_setcolorspace() so that, if we ever do
decide to alter the memory manager and allow for notifications, it
should be straight-forward to identify and replace the code.