Fix some kinds of type1 patterns
There was a problem when processing [/Pattern] colour spaces, we were
throwing an error, instead of treating it the same as a NULL colour
space array. Changing this resulted in 756 errors (about one quarter of
the current total) being resolved. In addition there were 576 files
which demonstrated progressions.

Also, if we get an error in pdf_pattern_create() and we have already
created the colour space, we must count it down. Move this into the
general exit condition.