Fix retention of the pdf14 compositor 'properly'
Nancy noticed the fact that gs_push_pdf14trans_device() uses 'retain'
to prevent the device disappearing. This is required for PostScript
(see GS commit dd2bd7b4ca1886a4f3c58afbdacbb93831c57e1b) but for other
interpreters this causes a memory leak.

Nancy had put a temporary work-around in place but, with the 'official'
method in the commit above we need to remove that work-around and use
the new means instead.