Fix memory leaks with Patterns
Several fixes here.

Firstly, rather to my surprise, the colour space management code in the
graphics library doesn't manage shading dictionaries attached to type
2 patterns. So we have to free the (graphics library created) shading
structure when the Pattern is freed ourselves. Fortunately we already
handle our own object management for type 1 patterns so we can use the
same approach.

The SCN/scn operator creates a pattern in a local scope variable which
it later passes to gs_setcolor() (we are defining a /Pattern colour
space and the actual pattern is the 'colour). We weren't counting down
the reference held by the local variable, so when it went out of scope
we ended up with the reference count being incorrect and the pattern
was never freed.

We were not freeing the Decode array for the various mesh shading types.
This change partially simplified the code for creating/freeing shadings
by dropping the 'type' because we can figure it out ourselves. We could
simplify the pdfi_shading_free code further.

Shamelessly stole Robin's fix in commit d181151975f0ee13162cc03865c0cc5f58931823
which calls gx_pattern_cache_free().