Fix SEGV with psdcmykog device and files with > 7 spots.
Example invocation:

 membin/gs -Ilib -sOutputFile=out.%d.psd -dMaxBitmap=10000
     -sDEVICE=psdcmykog -r600 -Z: -sDEFAULTPAPERSIZE=letter
     -dNOPAUSE -dBATCH -K2000000 -dClusterJob -dJOBSERVER
     ./tests_private/comparefiles/Bug693541.pdf

The PDF14 buffers are set up to have dev->color_info.num_components
planes. When these buffers are written into, components 0 to num_spots-1
inclusive are written. If num_spots > num_components, we overwrite
data.

Devices such as psdcmyk expand num_components to accomodate the
number of spots found on the page. Devices such as psdcmykog
do not; they stick with the number of spots they know about to
start with (i.e. 2).

Accordingly, if we run a file through with (say) 11 spots on the
page, with a psdcmykog device, the pdf14 device is setup with
num_comps = 7 (6+alpha), but the number of spots is taken as
the number of spots on the page (11), and thus we have memory
overwrites.

The fix attempted here is to set the number of spots to be
num_comps - number_of_standard_comps.