Fix SEGV seen in -dUseCIEColor testing.
The following test fails with a SEGV in overnight testing:

  tests_private/ps/ps3cet/30-07.PS.psdcmyk.300.1

I can reproduce the problem with this call:

 membin/gs -Ilib -sOutputFile=out.%d.psd -dUseCIEColor -dMaxBitmap=10000
 -sDEVICE=psdcmyk -r300 -Z: -sDEFAULTPAPERSIZE=letter -dNOPAUSE -dBATCH
 -K2000000 -dClusterJob -dJOBSERVER -dCETMODE %rom%Resource/Init/gs_cet.ps
 tests_private/ps/ps3cet/30-07.PS

(with a hacked version of Memento so that blocks aren't all put into
their own chunk).

The problem is that when a shading is built it contains a set of
params, one of which is a reference to the cie_joint_caches
structure. This structure is held in a chunk of memory allocated
from pgs->memory. This entire chunk of memory goes away when an
l2restore happens. The built shading, however, is referenced by
a PS object that can (and does) outlive this.

When a gc happens, it follows the pointer as part of the enumeration
and crashes due to it having been freed already.

The fix implemented here, is to add a new graphics lib entrypoint
that returns a reference to a copy of the current caches, allocated
in a given memory space; we use that to get a reference to the
caches in the same memory space (and hence with the same lifespan
as) the shading.