Implement the 'split_bogus_operator' function
Previously left as a FIXME. If we find a token that we don't understand
in a PDF file, this code attempts to split it into 2 (and only 2) PDF
operators/keywords, on the assumption that an end-of-line character has
been missed off.

Examples include "reendstream", "cmendstream", "QBT".

Previously we just threw away the corrupted token, which coudl lead to
a lot of errors being thrown.