Bug 701296: Fix Interpolation in Indexed spaces with UseFastColor
The get_color_handler logic was being called with pcs = Indexed, and
so doing pcs = pcs->base_space. Then it was checking that pcs == the
concrete space, seeing that it was, and deciding it could then use
the device colors direct. The problem is that pcs->base_space is a
greyscale space, so when running to (say) an rgb output, we have the
wrong color data.

We now check for the number of components being as we expect, and
for the space being one of our 'default' ones before choosing to
take the 'fast' route.