Permit extra widths on space glyphs (PDF Tw value) even for xyshow
For the benefit of the pdfi interpreter. With the PDF in PostScript
current interpreter, we apply Widths using a cshow procedure. The pdfi
interpreter instead uses the x_widths and y_widths arrays, intended for
use with the PostScript xyshow operator instead.

However, the existing code doesn't add the extra width to space glyphs
if we have TEXT_REPLACE_WIDTHS set (ie an xyshow).

Here we move the code for the additional width out of the 'else'
clause in show_move() so that we always have the value available. We
then apply the dx and/or dy values in both cases.

We also need to relax the 'invalid' checking which is performed at a
higher level in the code (gx_device_text_begin) because this combination
is explicitly barred there and returns an error.