Many, many font and text changes.
Point size is now handled as part of the text matrix, so all fonts are
scaled to a nominal size of 1. Thus we no longer need to pass the
font size to the font creation routines.

For type 3 fonts; if we cannot find a named glyph in the font's
CharProcs dictionary, try to find a glyph named /.notdef. If we find
that, then use it in place of the missing glyph.

We need to store the original PDF font dictionary in the type 3 font,
rather than the Resources dictionary, because the find_resource
machinery expects to be given a dictionary containing a /Resources key.
We needed the font dictionary anyway so this is no hardship. Pass the
font dictionary to pdfi_interpret_inner_content_stream() instead of
the Resources dictionary.

Correctly set the id and UID values in the font, so that the glyph
cache can identify glyphs from different fonts.

Fixes for the TL, T*, Tc, Td, TD, Ts, Tw operators

Add a check for a degenerate text matrix in the Tm code.

Implement the ' and " operators

Rework the Tj and TJ code; we now use the TEXT_REPLACE_WIDTHS bit of
the text enumerator, and set the x_widths and y_widths arrays, so that
we can (effectively) use an xyshow to implement the font /Widths array
when drawing text. Add Any Tc values as well. Implement the text
render matrix, include in it the font point size, text rise, text
horizontal scaling which allows us to implement those features. Save
and restore the CTM (and *only* the CTM and current point) round text
operations, update the text matrix with the current point after
drawing text.

Still to do; implement text rendering modes, fix a couple of files
which show problems when rendering.