First pass at text rendering modes
Unfortunately the only font type we support at present (type 3) is
specified as ignoring text rendering modes. Acrobat doesn't quite do
that in fact, and nor does Ghostscript, but the limited support offered
isn't enough to really test this.

So I expect this will need to change when we can properly test it.