Working on transparency in patterns and forms
Test file pattrans_simple.pdf
Test file tests_private/comparefiles/Bug691763.pdf

- Turn on page group code in pdfi_render_page()

- Added code to more closely follow the ps code.
GS_SMask() implementation was actually rendering the SMask, when it was just
supposed to save it for later.
Wrote pdfi_trans_set_mask() based on the code that was in GS_SMask, plus
a bunch of fixes.

- Added implementations of 'q' and 'Q'.  These need to push/pop the
the transparency state, while normal pdfi_grestore() and pdfi_gsave() do not.

- Implement client_data for pdfi
Right now, the only thing in there is the SMask, which will now be tracked
with gsave/grestore.  The postscript implementation has a bunch of other
stuff in its client_data that may also belong in ours?

- Refactor the Form code in pdfi_Do()
It needs to be called various ways for SMasks, forms, etc.  Big hairy mess.

- Fixed some issues in pdf_trans.c
It appears the names of the Knockout and Isolated keys are /K and /I.
Was passing in PDF14_BEGIN_TRANS_PAGE_GROUP instead of PDF14_BEGIN_TRANS_GROUP
for regular transparency groups (copy/paste error).