Fix cases where there is confusion over an ICCBased colour space, the value of /N and the actual profile.
This is to deal with Bug696690.pdf, where the profile is broken but
claims to be an RGB profile, while the value of /N is 1 (which is in
fact correct). At the same time we need to use the number of components
from the profile in preference to the value of /N when the profile is
valid (eg Bug 696120.pdf).

We achieve this by using the number of profile components to build an
ICCBased space, even when that differs from /N, and we return the
number of components in the profile. The caller then checks /N and the
returned value and, if they differ, attempts to set the colour space,
and a dummy colour value, in order to test out the profile. If that
succeeds then we use the space, if it fails then we fall back to
checking for an Alternate and using that, and if there is no Alternate
then we use the original /N value to select a device space.

NB if we attempt to use the Alternate for an ICCBased space we now
insist it must be a name, not an array, which prevents us trying to use
a /Indexed space as the Alternate to an ICCBased space.