Detect invalid switch to non-clip text rendering mode
Switching from a text rendering mode including any clip, to a text
rendering mode which does not include a clip is invalid, according to
the PDF Reference Manual. We (astonishingly) don't seem to have an
example of a PDF file which does this, but future proof with a warning
if we ever encounter one.