Fix Separation and DeviceN Alternate colour space generation
Bug #701369 "PB with Pantone color in pdfwrite"

When converting to a specified colour space, if we encounter a DeviceN
or Separation space, we leave the spot inks in place but change the
alternate space to the defined colour conversion space, and generate a
new tint transform which approximates the original colours.

However there was a logical error in the code, which only became
apparent if TransferFunctionInfo was set to something other than 'Apply'

Normally this doesn't happen, but commit
2e68cc460dbe349f68b81082ff7344db48eb4820 changed the behvaiour when
we are doing colour conversion to prevent us trying to apply a transfer
function to a SMask, which we can't do.

This then exposed the fact that we were not copying the concretized
samples to the client_colour structure, but we were reading them *back*
from the client_colour structure to set the low and high values in the
generated tint transform function.

The solution is to copy the components from the concretized colour
outside the if clause testing the TransferFunctionInfo.

This fix continues to work properly with the file from bug #700052