pdfwrite - improve CIDFont generation
Bug 701010 "Low Printer VM and gswinc32.exe crashing on postscript file with Korean text"

The problem in this case is that the PostScript uses a CIDFont with two
descendant fonts. PDF only permits a single descendant font in a CIDFont
so pdfwrite can't simply copy the font 'as is'.

Instead we create a type 0 font (CIDFont) for each descendant and switch
between them as required.

The problem here is that the code for retrieving an existing Type 0
CIDFont for a given descendant in the PostScript assumed that the name
of the CIDFont would be of the form /FontName-CMapName, and the font
we have here does not match that.

In fact, I'm not at all sure that this is *ever* the case, certainly
none of our test files seem to exercise this (bearing in mind that its
the Type 0 CIDFont generated by pdfwrite that we are checking against
and I can see no code that created a name of that form).

So what was happening was that every time the PostScript CIDFont
switched descendants, we generated a new Yype 0 font, along with all the
attendant structures, such as a ToUnicode CMap (256kB on its own). In
fact the first line of text in the file switches descendants 26 times,
leading to us creating something like 21 MB of extra font information.

We now extend the check so that if the name is not FontName-CMapName we
check the FontName and the CMap Name (which we do also store) separately
and if they match we reuse the existing Type 0 font.

The supplied test file now runs to completion, all 217 pages, with a
peak memory usage of ~22.8MB whereas it previously exhausted memory,
using ~2GB, at page 111. In addition the resulting PDF file is around
95% smaller and the performance is noticeably improved on this file.